its the final countdown do do dooo doo

+++ demo.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 OUR_DIR=$(pwd)
+DRY_RUN=0
+
+# Dry
+#DRY_RUN=1
+SPEED=0.01
+
 # Disable git's paging which can cause a hang
 export GIT_PAGER=cat
 export GALAXY_HOSTNAME="$(hostname -f)"
@@ -19,8 +25,8 @@ clear
 
 # Install ansible (move to pre-setup?)
 cd git-gat
-pwd
-ls
+#pwd
+#ls
 
 # Show what's there
 ansible-galaxy install -p roles galaxyproject.self_signed_certs > /dev/null 2>/dev/null
@@ -31,27 +37,41 @@ IFS=$'\n'
 for thing in $(cat .scripts/10-ansible-galaxy-script.txt); do
 	if [[ "$thing" == "git checkout"* ]]; then
 		# Checkout this commit
-		bash -c "$thing"
+		bash -c "$thing -q"
 		current_Commit=$(git show --format=%H | head -n 1)
 		echo "$(tput bold)Next step: $(git show --format="%s" | head -n 1 | sed 's/[^:]*: //g') $(tput sgr0)"
-		sleep 2
+		if (( DRY_RUN == 0 )); then
+			sleep 2
+		fi
 
+		# This will pretend to edit it in """"vim""""
+		fileChanged=$(git show --name-only | tail -n 1)
 		# Checkout the previous commit so we can show the diff properly
 		git checkout -q "$current_Commit^1"
-		# This will pretend to edit it in """"vim""""
+		p "nano $fileChanged"
 		python3 ${OUR_DIR}/diffplayer.py \
 			--diff <(git show "$current_Commit") \
 			--nosave \
-			--speed 0.1
+			--speed ${SPEED}
 		# And revert back to the commit we said we're on
 		git checkout -q "$current_Commit"
 	elif [[ "$thing" ==  "ansible-galaxy"* ]]; then
 		echo "The requirements have changed, we'll need to install the new ones."
-		pe "ansible-galaxy install -p roles -r requirements.yml"
+		if (( DRY_RUN == 1 )); then
+			p "ansible-galaxy install -p roles -r requirements.yml"
+		else
+			pe "ansible-galaxy install -p roles -r requirements.yml"
+		fi
 
 	elif [[ "$thing" ==  "ansible-playbook"* ]]; then
 		echo "Let's re-run the playbook"
 		p "ansible-playbook galaxy.yml -u ${USER}"
-		ansible-playbook galaxy.yml -u ${USER} -e "nginx_ssl_role=galaxyproject.self_signed_certs openssl_domains={{ certbot_domains }}" --vault-password-file ~/.vault-password.txt -i ~/.hosts
+		if (( DRY_RUN == 1 )); then
+			echo ansible-playbook galaxy.yml -u ${USER} -e "nginx_ssl_role=galaxyproject.self_signed_certs openssl_domains={{ certbot_domains }}" --vault-password-file ~/.vault-password.txt -i ~/.hosts
+		else
+			ansible-playbook galaxy.yml -u ${USER} -e "nginx_ssl_role=galaxyproject.self_signed_certs openssl_domains={{ certbot_domains }}" --vault-password-file ~/.vault-password.txt -i ~/.hosts
+		fi
+	else
+		echo "UNKNOWN $thing"
 	fi
 done
